Reduce debug logging verbosity
The debug logging in main.py and database/session.py has been reduced to improve performance and readability. In addition, the log level in logging_config.py has been changed from INFO to DEBUG. This will result in less clutter in the logs and make it easier to spot relevant information.